Add portage configurations
Portage is a highly extensible and organized package manager that is the
main feature of Gentoo Linux. As this controls most of the Gentoo
features, its configuration is highly useful for me. For others, I would
recommend to take inspiration from it, not outright copy it as it WILL
break their system.